migrated

+++ Yone/Database/afk_sql.py
@@ -3,13 +3,13 @@ import threading
 from datetime import datetime
 
 from Yone.Database import BASE, SESSION
-from sqlalchemy import Boolean, Column, Integer, UnicodeText, DateTime
+from sqlalchemy import Boolean, Column, BigInteger, UnicodeText, DateTime
 
 
 class AFK(BASE):
     __tablename__ = "afk_users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     is_afk = Column(Boolean)
     reason = Column(UnicodeText)
     time = Column(DateTime)
